requirements.txt
# Always use conda for scientific packages.
# Core Streamlit Application Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI Model APIs  
google-generativeai>=0.3.0
openai>=1.3.0

# File Processing
PyPDF2>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
python-docx>=0.8.11
docx2txt>=0.8

# Firebase
firebase-admin>=6.2.0 